update boqian's md file, refactor exam and lesson commands to be in separate folders
