+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -9,26 +9,30 @@ import java.util.regex.Pattern;
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CreateExamCommand;
-import seedu.address.logic.commands.CreateLessonCommand;
 import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.DeleteExamCommand;
-import seedu.address.logic.commands.DeleteLessonCommand;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.UpdateExamCommand;
-import seedu.address.logic.commands.UpdateLessonCommand;
-import seedu.address.logic.commands.ViewExamCommand;
-import seedu.address.logic.commands.ViewLessonCommand;
+import seedu.address.logic.commands.exam.CreateExamCommand;
+import seedu.address.logic.commands.exam.DeleteExamCommand;
+import seedu.address.logic.commands.exam.UpdateExamCommand;
+import seedu.address.logic.commands.exam.ViewExamCommand;
 import seedu.address.logic.commands.homework.CreateHomeworkCommand;
 import seedu.address.logic.commands.homework.DeleteHomeworkCommand;
 import seedu.address.logic.commands.homework.MarkHomeworkAsDoneCommand;
 import seedu.address.logic.commands.homework.MarkHomeworkAsUndoCommand;
 import seedu.address.logic.commands.homework.UpdateHomeworkCommand;
 import seedu.address.logic.commands.homework.ViewHomeworkCommand;
+import seedu.address.logic.commands.lesson.CreateLessonCommand;
+import seedu.address.logic.commands.lesson.DeleteLessonCommand;
+import seedu.address.logic.commands.lesson.UpdateLessonCommand;
+import seedu.address.logic.commands.lesson.ViewLessonCommand;
+import seedu.address.logic.parser.exam.CreateExamCommandParser;
+import seedu.address.logic.parser.exam.DeleteExamCommandParser;
+import seedu.address.logic.parser.exam.UpdateExamCommandParser;
+import seedu.address.logic.parser.exam.ViewExamCommandParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.logic.parser.homework.CreateHomeworkCommandParser;
 import seedu.address.logic.parser.homework.DeleteHomeworkCommandParser;
@@ -36,6 +40,10 @@ import seedu.address.logic.parser.homework.MarkHomeworkAsDoneCommandParser;
 import seedu.address.logic.parser.homework.MarkHomeworkAsUndoCommandParser;
 import seedu.address.logic.parser.homework.UpdateHomeworkCommandParser;
 import seedu.address.logic.parser.homework.ViewHomeworkCommandParser;
+import seedu.address.logic.parser.lesson.CreateLessonCommandParser;
+import seedu.address.logic.parser.lesson.DeleteLessonCommandParser;
+import seedu.address.logic.parser.lesson.UpdateLessonCommandParser;
+import seedu.address.logic.parser.lesson.ViewLessonCommandParser;
 
 /**
  * Parses user input.
