--- src/main/java/seedu/address/logic/parser/DeleteLessonCommandParser.java
@@ -1,62 +1,0 @@
-package seedu.address.logic.parser;
-
-import static java.util.Objects.requireNonNull;
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_INDEX;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Stream;
-
-import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteLessonCommand;
-import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.student.NamePredicate;
-
-/**
- * Parses input arguments and creates a new DeleteLessonCommand object
- */
-public class DeleteLessonCommandParser implements Parser<DeleteLessonCommand> {
-    private List<String> inputNames = new ArrayList<>();
-    /**
-     * Parses the given {@code String} of arguments in the context of the DeleteLessonCommand
-     * and returns a CreateHomeworkCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public DeleteLessonCommand parse(String args) throws ParseException {
-        requireNonNull(args);
-
-        ArgumentMultimap argMultimap =
-            ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_INDEX);
-
-        if (!arePrefixesPresent(argMultimap, PREFIX_NAME, PREFIX_INDEX)
-            || !argMultimap.getPreamble().isEmpty()) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                DeleteLessonCommand.MESSAGE_USAGE));
-        }
-
-        List<String> nameKeywords = argMultimap.getAllValues(PREFIX_NAME);
-        // for all the names, trim the name and only take the first word
-        for (int i = 0; i < nameKeywords.size(); i++) {
-            String name = nameKeywords.get(i);
-            name = name.trim();
-            //            int spaceIndex = name.indexOf(" ");
-            //            if (spaceIndex != -1) {
-            //                name = name.substring(0, spaceIndex);
-            //            }
-            nameKeywords.set(i, name);
-        }
-        inputNames = nameKeywords;
-
-        Index index = ParserUtil.parseIndex(argMultimap.getValue(PREFIX_INDEX).get());
-        return new DeleteLessonCommand(inputNames, new NamePredicate(nameKeywords), index);
-    }
-
-    /**
-     * Returns true if all prefixes are present in the given {@code ArgumentMultimap}.
-     */
-    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
-        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
-    }
-}
