+++ src/main/java/seedu/address/logic/parser/exam/CreateExamCommandParser.java
@@ -1,9 +1,9 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.exam;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ENDTIME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_LESSON;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_EXAM;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_STARTTIME;
 
@@ -12,40 +12,41 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.CreateLessonCommand;
+import seedu.address.logic.commands.exam.CreateExamCommand;
+import seedu.address.logic.parser.ArgumentMultimap;
+import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
+import seedu.address.logic.parser.Prefix;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.student.NamePredicate;
 
 /**
- * Parses input arguments and creates a new CreateHomeworkCommand object
+ * Parses input arguments and creates a new CreateHomeworkExam object
  */
-public class CreateLessonCommandParser implements Parser<CreateLessonCommand> {
+public class CreateExamCommandParser implements Parser<CreateExamCommand> {
     private List<String> names = new ArrayList<>();
     /**
-     * Parses the given {@code String} of arguments in the context of the CreateHomeworkCommand
-     * and returns a CreateHomeworkCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the CreateExamCommand
+     * and returns a CreateExamCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public CreateLessonCommand parse(String args) throws ParseException {
+    public CreateExamCommand parse(String args) throws ParseException {
         requireNonNull(args);
 
         ArgumentMultimap argMultimap =
-            ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_LESSON, PREFIX_STARTTIME, PREFIX_ENDTIME);
+                ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_EXAM, PREFIX_STARTTIME, PREFIX_ENDTIME);
 
-        if (!arePrefixesPresent(argMultimap, PREFIX_NAME, PREFIX_LESSON, PREFIX_STARTTIME, PREFIX_ENDTIME)
-            || !argMultimap.getPreamble().isEmpty()) {
+        if (!arePrefixesPresent(argMultimap, PREFIX_NAME, PREFIX_EXAM, PREFIX_STARTTIME, PREFIX_ENDTIME)
+                || !argMultimap.getPreamble().isEmpty()) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                CreateLessonCommand.MESSAGE_USAGE));
+                    CreateExamCommand.MESSAGE_USAGE));
         }
 
-        String lessonName = argMultimap.getValue(PREFIX_LESSON).get();
+        String examDescription = argMultimap.getValue(PREFIX_EXAM).get();
         LocalDateTime startTime = ParserUtil.parseDeadline(argMultimap.getValue(PREFIX_STARTTIME).get());
         LocalDateTime endTime = ParserUtil.parseDeadline(argMultimap.getValue(PREFIX_ENDTIME).get());
         List<String> nameKeywords = argMultimap.getAllValues(PREFIX_NAME);
-
-        if (endTime.isBefore(startTime)) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, CreateLessonCommand.MESSAGE_DATE));
-        }
 
         // for all the names, trim the name and only take the first word
         for (int i = 0; i < nameKeywords.size(); i++) {
@@ -59,8 +60,8 @@ public class CreateLessonCommandParser implements Parser<CreateLessonCommand> {
         }
         names = nameKeywords;
 
-        return new CreateLessonCommand(names, new NamePredicate(nameKeywords),
-            lessonName, startTime, endTime);
+        return new CreateExamCommand(names, new NamePredicate(nameKeywords),
+                examDescription, startTime, endTime);
     }
 
     /**
