+++ src/main/java/seedu/address/logic/parser/exam/ViewExamCommandParser.java
@@ -1,11 +1,11 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.exam;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_DATE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_DONE;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_EXAM;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_SUBJECT;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_STUDENTS;
 
 import java.time.LocalDate;
@@ -13,41 +13,45 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Predicate;
 
-import seedu.address.logic.commands.ViewLessonCommand;
+import seedu.address.logic.commands.exam.ViewExamCommand;
+import seedu.address.logic.parser.ArgumentMultimap;
+import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.student.Lesson;
-import seedu.address.model.student.LessonBelongsToDatePredicate;
-import seedu.address.model.student.LessonDonePredicate;
-import seedu.address.model.student.LessonSubjectPredicate;
+import seedu.address.model.student.Exam;
+import seedu.address.model.student.ExamDatePredicate;
+import seedu.address.model.student.ExamDonePredicate;
+import seedu.address.model.student.ExamPredicate;
 import seedu.address.model.student.NamePredicate;
 import seedu.address.model.student.Student;
 
 /**
  * Parses input arguments and creates a new ViewHomeworkCommand object
  */
-public class ViewLessonCommandParser implements Parser<ViewLessonCommand> {
+public class ViewExamCommandParser implements Parser<ViewExamCommand> {
     /**
      * Parses the given {@code String} of arguments in the context of the ViewLessonCommand
      * and returns a ViewLessonCommand object for execution.
      * @param args the user input to be parsed into a ViewLessonCommand object.
-     * @return a ViewLessonCommand object.
+     * @return a ViewExamCommand object.
      */
-    public ViewLessonCommand parse(String args) throws ParseException {
+    public ViewExamCommand parse(String args) throws ParseException {
         requireNonNull(args);
 
-        ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_DATE, PREFIX_SUBJECT,
+        ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_DATE, PREFIX_EXAM,
             PREFIX_DONE);
 
         if (!argMultimap.getPreamble().isEmpty()) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                ViewLessonCommand.MESSAGE_USAGE));
+                ViewExamCommand.MESSAGE_USAGE));
         }
 
         Predicate<Student> namePredicate;
-        Predicate<Lesson> subjectPredicate = lesson -> true;
-        Predicate<Lesson> donePredicate = lesson -> true;
-        List<String> nameList = new ArrayList<>();
+        Predicate<Exam> examPredicate = exam -> true;
+        Predicate<Exam> donePredicate = exam -> true;
         boolean defaultPredicateFlag;
+        List<String> nameList = new ArrayList<>();
 
         // If name is present, create a predicate to filter by name
         if (argMultimap.getValue(PREFIX_NAME).isPresent()) {
@@ -71,33 +75,25 @@ public class ViewLessonCommandParser implements Parser<ViewLessonCommand> {
             defaultPredicateFlag = true;
         }
 
-        if (argMultimap.getValue(PREFIX_SUBJECT).isPresent()) {
-            String subject = argMultimap.getValue(PREFIX_SUBJECT).get();
-            subjectPredicate = new LessonSubjectPredicate(subject);
+        if (argMultimap.getValue(PREFIX_EXAM).isPresent()) {
+            String exam = argMultimap.getValue(PREFIX_EXAM).get();
+            examPredicate = new ExamPredicate(exam);
         }
 
         if (argMultimap.getValue(PREFIX_DONE).isPresent()) {
             String done = argMultimap.getValue(PREFIX_DONE).get();
-            donePredicate = new LessonDonePredicate(done);
+            donePredicate = new ExamDonePredicate(done);
         }
 
-        // If date is present, create a predicate to filter by status
+        // If date is present, create a predicate to filter by date
         if (argMultimap.getValue(PREFIX_DATE).isPresent()) {
             String date = argMultimap.getValue(PREFIX_DATE).get();
             LocalDate targetDate = ParserUtil.parseDate(date);
-            LessonBelongsToDatePredicate datePredicate = new LessonBelongsToDatePredicate(targetDate);
-            return new ViewLessonCommand(nameList, namePredicate, datePredicate, subjectPredicate, donePredicate,
+            ExamDatePredicate datePredicate = new ExamDatePredicate(targetDate);
+            return new ViewExamCommand(nameList, namePredicate, datePredicate, examPredicate, donePredicate,
                 defaultPredicateFlag);
         } else {
-            return new ViewLessonCommand(nameList, namePredicate, subjectPredicate, donePredicate,
-                defaultPredicateFlag);
+            return new ViewExamCommand(nameList, namePredicate, examPredicate, donePredicate, defaultPredicateFlag);
         }
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-            || (other instanceof ViewLessonCommandParser // instanceof handles nulls
-            && this.equals((ViewLessonCommandParser) other)); // state check
     }
 }
